requirements.txt
accelerate==0.28.0
albumentations==1.4.2
einops==0.8.1
kornia==0.7.2
mamba_ssm==2.2.2
numpy==2.2.4
opencv_python==4.10.0.84
opencv_python_headless==4.9.0.80
Pillow==11.2.1
skimage==0.0
thop==0.1.1.post2209072238
torch==2.2.0
torchmetrics==1.3.2
torchvision==0.17.0
tqdm==4.66.2
yacs==0.1.8
